chore(packages/notifications): move notifications to packages
