+++ packages/components/notifications/content-notification/index.js
@@ -2,5 +2,5 @@
 // if this file is missing, and you have a `module` or `main` that points to a non-existing file
 // (ie, a bundle that hasn't been built yet) then jest will fail if the bundle is not yet built.
 // all apps should export all their named exports from their root index.js
-export { default as ContentNotification } from './content-notification';
-export * from './content-notification';
+export { default } from './src';
+export * from './src';
