+++ packages/components/notifications/content-notification/src/content-notification.js
