+++ packages/components/notifications/content-notification/src/content-notification.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ContentNotification } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 export const routePath = '/content-notification';
