+++ packages/components/notifications/content-notification/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './content-notification';
+
+export { default as version } from './version';
