+++ packages/components/notifications/content-notification/src/version.js
